refactor: replace passed type with a string

+++ app/src/main/java/com/immortalidiot/wishescompose/presentation/screens/DayWishGeneratorScreen.kt
@@ -46,7 +46,7 @@ fun DayWishGeneratorScreen(
             screenViewModel.changeNumberEmojis(changedEmojis)
         },
         onGenerateClick = {
-            screenViewModel.generateDayWishAndCopy(uiState.emojis.toInt())
+            screenViewModel.generateDayWishAndCopy(uiState.emojis)
         },
         onBackButton = {
             navHostController.popBackStack(
